add base class for pickups, move coin and money stuff into pickups folder
