--- Maze_Shooter/Assets/Scripts/Money.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ab36720d4cfd85046b8954fc2927eafb
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
