+++ Maze_Shooter/Assets/Scripts/Pickups/MoneySpawner.cs
