+++ Maze_Shooter/Assets/Scripts/Pickups/Pickup.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1be41b868472dde4a80d8a1376f361ae
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
