+++ Maze_Shooter/Assets/Scripts/Pickups/Wallet.cs
